% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fit Linear Regression Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{An object of class \code{"formula"}: a symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}
}
\value{
An object of class \code{"linearR"} containing the model fit results.
}
\description{
The `my_lm` function fits a linear regression model using the normal equations method.
It returns an object of class `"linearR"` containing the model fit results, including
coefficients, standard errors, t-statistics, p-values, residuals, fitted values, and
various summary statistics.
}
\details{
The function uses the normal equations method to estimate the coefficients:
\deqn{\hat{\beta} = (X^\top X)^{-1} X^\top y}
where \eqn{X} is the design matrix and \eqn{y} is the response vector.
}
\examples{
data(mtcars)
fit <- my_lm(mpg ~ wt + cyl, data = mtcars)
summary(fit)
}
