% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{print_summary_linearR}
\alias{print_summary_linearR}
\title{Print Method for summary_linearR Objects}
\usage{
print_summary_linearR(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary_linearR"}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints the summary of a `linearR` object generated by \code{\link{summary_linearR}}.
}
\examples{
data(mtcars)
fit <- my_lm(mpg ~ wt + cyl, data = mtcars)
summary_fit <- summary_linearR(fit)
print_summary_linearR(summary_fit)
}
